+++ R/parsedate-package.r
@@ -191,6 +191,10 @@ parse_iso_8601 <- function(dates, default_tz = "UTC") {
 
 parse_iso_parts <- function(mm, default_tz) {
 
+  # Ensure that all fractions use periods rather than commas and are numeric
+  mm$frac <- as.numeric(sub(",", ".", mm$frac))
+  mm$sec <- as.numeric(sub(",", ".", mm$sec))
+
   num <- nrow(mm)
 
   ## Date first ----
@@ -229,27 +233,25 @@ parse_iso_parts <- function(mm, default_tz) {
   tm <- mm$min != ""
   date[tm] <- date[tm] + minutes(mm$min[tm])
 
-  ts <- mm$sec != ""
+  ts <- !is.na(mm$sec)
   date[ts] <- date[ts] + seconds(mm$sec[ts])
 
   ## Fractional time ----
 
-  frac <- as.numeric(sub(",", ".", mm$frac))
-
-  tfs <- !is.na(frac) & mm$sec != ""
+  tfs <- !is.na(mm$frac) & !is.na(mm$sec)
   # only supporting up to millisecond resolution, rounding subsequent digits
-  date[tfs] <- date[tfs] + milliseconds(round(frac[tfs] * 1000))
+  date[tfs] <- date[tfs] + milliseconds(round(mm$frac[tfs] * 1000))
 
-  tfm <- !is.na(frac) & mm$sec == "" & mm$min != ""
-  sec <- trunc(frac[tfm] * 60)
+  tfm <- !is.na(mm$frac) & is.na(mm$sec) & mm$min != ""
+  sec <- trunc(mm$frac[tfm] * 60)
   # only supporting up to millisecond resolution, rounding subsequent digits
-  mil <- round((frac[tfm] * 60 - sec) * 1000)
+  mil <- round((mm$frac[tfm] * 60 - sec) * 1000)
   date[tfm] <- date[tfm] + seconds(sec) + milliseconds(mil)
 
-  tfh <- !is.na(frac) & mm$sec == "" & mm$min == ""
-  min <- trunc(frac[tfh] * 60)
-  sec <- trunc((frac[tfh] * 60 - min) * 60)
-  mil <- round((((frac[tfh] * 60) - min) * 60 - sec) * 1000)
+  tfh <- !is.na(mm$frac) & is.na(mm$sec) & mm$min == ""
+  min <- trunc(mm$frac[tfh] * 60)
+  sec <- trunc((mm$frac[tfh] * 60 - min) * 60)
+  mil <- round((((mm$frac[tfh] * 60) - min) * 60 - sec) * 1000)
   date[tfh] <- date[tfh] + minutes(min) + seconds(sec) + milliseconds(mil)
 
   ## Time zone ----
@@ -303,8 +305,8 @@ iso_regex <- paste0(
             "(?:(?<colon>:?)(?<min>[0-5]\\d))?|24\\:?00)",
            # the fraction after the minute
            "(?<frac>[\\.,]\\d+(?!:))?)?",
-    # the colon after the minute and the second
-    "(?:\\g{colon}(?<sec>[0-5]\\d)(?:[\\.,]\\d+)?)?",
+    # the colon after the minute and the second with its optional fraction
+    "(?:\\g{colon}(?<sec>[0-5]\\d(?:[\\.,]\\d+)?))?",
     # the timezone as "Z" or "+-hh" or "+-hh:mm" or "+-hhmm"
     "(?<tz>[zZ]|(?<tzpm>[\\+-])",
      "(?<tzhour>[01]\\d|2[0-3]):?(?<tzmin>[0-5]\\d)?)?)?)?$"